Move TestTextMessage to tests
